Move SmartButtonGroup to widgets. FontColorChooser extends ColorChooser
